fix(sonar): resolve blocker, bug, critical and minor issues
- BLOCKER: VortexGridTui.move/jumpRow/jumpCol always returned true;
  changed to void, updated switch arms to use unnamed patterns (_)
- BUG: PatchedData record — suppress S6218 for array-field records
- CRITICAL: reduce cognitive complexity in CsvImporter, CsvExporter,
  PatchedEncodingEncoder, ScanIterator, LazyGridSource via extraction
- MAJOR: remove unused ctx param in PatchedEncodingEncoder.Encoder,
  remove redundant long/double/float casts, fix loop counter declarations,
  suppress S106 on CLI commands (intentional System.out/err usage),
  rephrase VortexWriter comment that triggered false-positive S125
- MINOR: unnamed patterns (_) in catch/switch/TWR, {link} → [...] in
  Javadoc across array types, isZero() in zigzag tests, S4929 suppress
  on WindowsTerminal.ConsoleInputStream

+++ reader/src/main/java/io/github/dfa1/vortex/reader/decode/AlpEncodingDecoder.java
@@ -92,7 +92,7 @@ public final class AlpEncodingDecoder implements EncodingDecoder {
                 return new LazyAlpDoubleArray(ctx.dtype(), n, src, df, de);
             }
             // broadcast without patches: decode single encoded value → constant
-            double v = (double) src.getAtIndex(PTypeIO.LE_LONG, 0) * df * de;
+            double v = src.getAtIndex(PTypeIO.LE_LONG, 0) * df * de;
             return new LazyConstantDoubleArray(ctx.dtype(), n, v);
         }
 
@@ -122,7 +122,7 @@ public final class AlpEncodingDecoder implements EncodingDecoder {
                 return new LazyAlpFloatArray(ctx.dtype(), n, src, df, de);
             }
             // broadcast without patches: decode single encoded value → constant
-            float v = (float) src.getAtIndex(PTypeIO.LE_INT, 0) * df * de;
+            float v = src.getAtIndex(PTypeIO.LE_INT, 0) * df * de;
             return new LazyConstantFloatArray(ctx.dtype(), n, v);
         }
